==> 02-parse-annotations.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import datetime
import pickle
import yaml
import pandas as pd

# Load annotation data

annotation_data = pd.read_csv('./data/annotations.csv', low_memory=False)

list(enumerate(annotation_data.columns))

annotation_data.iloc[:, :12].sample(2).T

annotation_data['dt'] = annotation_data.apply(
    lambda x: f"{x.year}-{x.week}-{x.day + 1}",
    axis=1
).map(lambda x: datetime.datetime.strptime(x, '%Y-%W-%u'))

annotation_data['date_slno'] = (annotation_data.dt - annotation_data.dt.min()).dt.days

annotation_data = annotation_data.sort_values('audio_filename')
file_list = annotation_data['audio_filename'].unique().tolist()
len(file_list)

# Train/validate/test dataset exploration

annotation_data.shape

annotation_data.audio_filename.drop_duplicates().shape

(annotation_data
 .loc[:, ['split', 'audio_filename']]
 .drop_duplicates()
).shape

(annotation_data
 .loc[:, ['split', 'audio_filename', 'borough', 'block', 'latitude', 'longitude', 'year', 'week', 'day', 'hour', 'dt']]
 .drop_duplicates()
).shape

tmp = (annotation_data
       .loc[:, ['split', 'audio_filename', 'borough', 'block', 'latitude', 'longitude', 'year', 'week', 'day', 'hour', 'dt']]
       .drop_duplicates())

tmp.loc[:, ['split', 'audio_filename']].split.value_counts()

tmp.loc[lambda x: x.split == 'train'].borough.drop_duplicates().tolist()
tmp.loc[lambda x: x.split == 'validate'].borough.drop_duplicates().tolist()
tmp.loc[lambda x: x.split == 'test'].borough.drop_duplicates().tolist()

train_blocks = set(tmp.loc[lambda x: x.split == 'train'].block.tolist())
validate_blocks = set(tmp.loc[lambda x: x.split == 'validate'].block.tolist())
test_blocks = set(tmp.loc[lambda x: x.split == 'test'].block.tolist())

len(validate_blocks - train_blocks)
len(test_blocks - train_blocks)

tmp.loc[lambda x: x.split == 'train'].dt.min(), tmp.loc[lambda x: x.split == 'train'].dt.max()
tmp.loc[lambda x: x.split == 'validate'].dt.min(), tmp.loc[lambda x: x.split == 'validate'].dt.max()
tmp.loc[lambda x: x.split == 'test'].dt.min(), tmp.loc[lambda x: x.split == 'test'].dt.max()

tmp.loc[lambda x: x.split == 'train'].latitude.min(), tmp.loc[lambda x: x.split == 'train'].latitude.max()
tmp.loc[lambda x: x.split == 'validate'].latitude.min(), tmp.loc[lambda x: x.split == 'validate'].latitude.max()
tmp.loc[lambda x: x.split == 'test'].latitude.min(), tmp.loc[lambda x: x.split == 'test'].latitude.max()

tmp.loc[lambda x: x.split == 'train'].longitude.min(), tmp.loc[lambda x: x.split == 'train'].longitude.max()
tmp.loc[lambda x: x.split == 'validate'].longitude.min(), tmp.loc[lambda x: x.split == 'validate'].longitude.max()
tmp.loc[lambda x: x.split == 'test'].longitude.min(), tmp.loc[lambda x: x.split == 'test'].longitude.max()

# Parse taxonomy

with open('./data/dcase-ust-taxonomy.yaml', 'r') as f:
    taxonomy = yaml.load(f, Loader=yaml.Loader)

full_fine_target_labels = ["{}-{}_{}".format(coarse_id, fine_id, fine_label)
                           for coarse_id, fine_dict in taxonomy['fine'].items()
                           for fine_id, fine_label in fine_dict.items()]

fine_target_labels = [x
                      for x in full_fine_target_labels
                      if x.split('_')[0].split('-')[1] != 'X']

coarse_target_labels = ["_".join([str(k), v])
                        for k, v in taxonomy['coarse'].items()]

full_fine_target_labels
fine_target_labels
coarse_target_labels

tmp1 = pd.DataFrame.from_records([
    (key1, key2, val2)
    for key1, val1 in taxonomy['fine'].items()
    for key2, val2 in val1.items()
], columns=['coarse_id', 'fine_id', 'fine'])
tmp2 = pd.DataFrame.from_records([
    (key, val)
    for key, val in taxonomy['coarse'].items()
], columns=['coarse_id', 'coarse'])
taxonomy_df = pd.merge(tmp1, tmp2, on='coarse_id', how='inner')
taxonomy_df = taxonomy_df.astype('str')
taxonomy_df['coarse'] = taxonomy_df.coarse_id + '_' + taxonomy_df.coarse
taxonomy_df['fine'] = taxonomy_df.coarse_id + '-' + taxonomy_df.fine_id + '_' + taxonomy_df.fine

taxonomy_df

# Separate train and valid data, only zooniverse platform

train_data = (annotation_data.copy()
              .loc[lambda x: x.split == 'train']
              .loc[lambda x: x.annotator_id > 0]
              .set_index('audio_filename')
              .filter(regex='.*_presence$', axis='columns')
              .rename(columns=lambda x: x[:-9])
              .astype('int64'))

(train_data.reset_index()
 .assign(cnt=1).groupby('audio_filename')['cnt'].sum()
 .value_counts()
)

coarse_train = train_data.loc[:, sorted(taxonomy_df.coarse.unique())]
fine_train = train_data.loc[:, sorted(taxonomy_df.fine.tolist())]

coarse_train.shape, fine_train.shape, len(set(coarse_train.index.tolist()))

---

valid_data = (annotation_data.copy()
              .loc[lambda x: x.split == 'validate']
              .loc[lambda x: x.annotator_id > 0]
              .set_index('audio_filename')
              .filter(regex='.*_presence$', axis='columns')
              .rename(columns=lambda x: x[:-9])
              .astype('int64'))

(valid_data.reset_index()
 .assign(cnt=1).groupby('audio_filename')['cnt'].sum()
 .value_counts()
)

coarse_valid = valid_data.loc[:, sorted(taxonomy_df.coarse.unique())]
fine_valid = valid_data.loc[:, sorted(taxonomy_df.fine.tolist())]

coarse_valid.shape, fine_valid.shape, len(set(fine_valid.index.tolist()))

coarse_train.sort_index(inplace=True)
coarse_valid.sort_index(inplace=True)
fine_train.sort_index(inplace=True)
fine_valid.sort_index(inplace=True)

coarse_train.head(2)
coarse_valid.head(2)
fine_train.head(2)
fine_valid.head(2)

---

metadata_cols = [
 'sensor_id',
 'audio_filename',
 'borough',
 'block',
 'latitude',
 'longitude',
 'year',
 'week',
 'day',
 'hour',
 'dt',
 'date_slno']

train_metadata = (annotation_data.copy()
                  .loc[lambda x: x.split == 'train']
                  .loc[:, metadata_cols]
                  .drop_duplicates()
                  .set_index('audio_filename'))
valid_metadata = (annotation_data.copy()
                  .loc[lambda x: x.split == 'validate']
                  .loc[:, metadata_cols]
                  .drop_duplicates()
                  .set_index('audio_filename'))

train_metadata.shape, valid_metadata.shape

train_metadata.head(2)
valid_metadata.head(2)

# Separate train and valid data, only ground truth

train_data_gt = (annotation_data.copy()
                 .loc[lambda x: x.split == 'train']
                 .loc[lambda x: x.annotator_id == 0]
                 .set_index('audio_filename')
                 .filter(regex='.*_presence$', axis='columns')
                 .rename(columns=lambda x: x[:-9])
                 .astype('int64'))

(train_data_gt.reset_index()
 .assign(cnt=1).groupby('audio_filename')['cnt'].sum()
 .value_counts()
)

coarse_train_gt = train_data_gt.loc[:, sorted(taxonomy_df.coarse.unique())]
fine_train_gt = train_data_gt.loc[:, sorted(taxonomy_df.fine.tolist())]

coarse_train_gt.shape, fine_train_gt.shape, len(set(coarse_train_gt.index.tolist()))

valid_data_gt = (annotation_data.copy()
                 .loc[lambda x: x.split == 'validate']
                 .loc[lambda x: x.annotator_id == 0]
                 .set_index('audio_filename')
                 .filter(regex='.*_presence$', axis='columns')
                 .rename(columns=lambda x: x[:-9])
                 .astype('int64'))

(valid_data_gt.reset_index()
 .assign(cnt=1).groupby('audio_filename')['cnt'].sum()
 .value_counts()
)

coarse_valid_gt = valid_data_gt.loc[:, sorted(taxonomy_df.coarse.unique())]
fine_valid_gt = valid_data_gt.loc[:, sorted(taxonomy_df.fine.tolist())]

coarse_valid_gt.shape, fine_valid_gt.shape, len(set(coarse_valid_gt.index.tolist()))

coarse_train_gt.sort_index(inplace=True)
coarse_valid_gt.sort_index(inplace=True)
fine_train_gt.sort_index(inplace=True)
fine_valid_gt.sort_index(inplace=True)

coarse_train_gt.head(2)
coarse_valid_gt.head(2)
fine_train_gt.head(2)
fine_valid_gt.head(2)

---

train_metadata_gt = (annotation_data.copy()
                     .loc[lambda x: x.split == 'train']
                     .loc[lambda x: x.annotator_id == 0]
                     .loc[:, metadata_cols]
                     .drop_duplicates()
                     .set_index('audio_filename'))
valid_metadata_gt = (annotation_data.copy()
                     .loc[lambda x: x.split == 'validate']
                     .loc[lambda x: x.annotator_id == 0]
                     .loc[:, metadata_cols]
                     .drop_duplicates()
                     .set_index('audio_filename'))

train_metadata_gt.shape, valid_metadata_gt.shape

train_metadata_gt.head(2)
valid_metadata_gt.head(2)

# Save metadata

with open('./data/metadata.pkl', 'wb') as f:
    pickle.dump({
        'coarse_train': coarse_train,
        'coarse_valid': coarse_valid,
        'fine_train': fine_train,
        'fine_valid': fine_valid,
        'train_metadata': train_metadata,
        'valid_metadata': valid_metadata,
        'coarse_train_gt': coarse_train_gt,
        'coarse_valid_gt': coarse_valid_gt,
        'fine_train_gt': fine_train_gt,
        'fine_valid_gt': fine_valid_gt,
        'train_metadata_gt': train_metadata_gt,
        'valid_metadata_gt': valid_metadata_gt,
        'taxonomy_df': taxonomy_df
    }, f)
