==> 08-mobilenet-imagenet-metadata-all.ipynb <==
# Load libraries

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import random
import pickle
import numpy as np
import pandas as pd

import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader, Dataset

import torchvision.models
from torchvision import transforms

# https://pytorch.org/docs/stable/notes/randomness.html

seed = 20200701
random.seed(seed)
torch.manual_seed(seed)
np.random.seed(seed)

torch.backends.cudnn.deterministic = True
torch.backends.cudnn.benchmark = False

from torch.utils.tensorboard import SummaryWriter

this_hparams = {
    'slno': 8,
    'weights': 'imagenet',
    'dataaug': 'mixup',
    'comments': 'metadata-all'
}

import shutil
this_filename = f"{this_hparams['slno']}-{this_hparams['weights']}-{this_hparams['dataaug']}-{this_hparams['comments']}"
this_log_dir = f"./tfboard_out/{this_filename}"
shutil.rmtree(this_log_dir, ignore_errors=True)
tbwriter = SummaryWriter(this_log_dir)

this_filename

# Prepare the data

def prepare_data(df, unknown_to_known):
    df = df.reset_index()
    df['slno'] = df.assign(slno=1).groupby('audio_filename')['slno'].cumsum()
    df.set_index(['audio_filename', 'slno'], inplace=True)

    df_unknown = df.copy().loc[:, list(unknown_to_known.keys())]
    df.drop(columns=list(unknown_to_known.keys()), inplace=True)

    y_mask = df.copy()
    y_mask.loc[:, :] = 1
    for unknown, known in unknown_to_known.items():
        y_mask.loc[
            df_unknown[unknown] > 0.5,
            known
        ] = 0

    df = df.swaplevel(i=1, j=0, axis=0).sort_index()

    y_mask = y_mask.swaplevel(i=1, j=0, axis=0).sort_index()

    y = np.concatenate([
        df.loc[[1], :].values[..., np.newaxis],
        df.loc[[2], :].values[..., np.newaxis],
        df.loc[[3], :].values[..., np.newaxis]
    ], axis=2)

    y_mask = np.concatenate([
        y_mask.loc[[1], :].values[..., np.newaxis],
        y_mask.loc[[2], :].values[..., np.newaxis],
        y_mask.loc[[3], :].values[..., np.newaxis]
    ], axis=2)

    X = np.concatenate([
        np.expand_dims(np.load('./data/logmelspec/{}.npy'.format(x)).T[:635, :], axis=0)
        for x in df.loc[[1], :].reset_index(1).audio_filename.tolist()])
    X = np.expand_dims(X, axis=1)

    return X, y, y_mask


with open('./data/metadata.pkl', 'rb') as f:
    metadata = pickle.load(f)

unknown_to_known = (
    pd.merge(metadata['taxonomy_df'].loc[lambda x: x.fine_id == 'X', ['fine', 'coarse']],
             metadata['taxonomy_df'].loc[lambda x: x.fine_id != 'X', ['fine', 'coarse']],
             on='coarse', how='inner')
    .drop(columns='coarse')
    .groupby('fine_x')['fine_y']
    .apply(lambda x: list(x)).to_dict())
known_labels = metadata['taxonomy_df'].loc[lambda x: x.fine_id != 'X'].fine.tolist()

train_df = pd.concat([metadata['coarse_train'], metadata['fine_train']], axis=1, sort=True)
valid_df = pd.concat([metadata['coarse_valid'], metadata['fine_valid']], axis=1, sort=True)

train_X, train_y, train_y_mask = prepare_data(train_df, unknown_to_known)
valid_X, valid_y, valid_y_mask = prepare_data(valid_df, unknown_to_known)

---

cols_to_select = ['borough', 'latitude', 'longitude', 'week', 'day', 'hour', 'date_slno']
train_metadata = metadata['train_metadata'].loc[:, cols_to_select]
valid_metadata = metadata['valid_metadata'].loc[:, cols_to_select]

for x in ['borough', 'day', 'hour']:
    train_metadata[x] = train_metadata[x].astype('str')
    valid_metadata[x] = valid_metadata[x].astype('str')

train_metadata = pd.get_dummies(train_metadata)

valid_metadata = pd.get_dummies(valid_metadata)

valid_metadata['borough_4'] = 0

valid_metadata = valid_metadata.loc[:, train_metadata.columns.tolist()]

lat_min = train_metadata.latitude.min()
lat_max = train_metadata.latitude.max()
train_metadata.latitude = (train_metadata.latitude - lat_min) / (lat_max - lat_min)
valid_metadata.latitude = (valid_metadata.latitude - lat_min) / (lat_max - lat_min)

lon_min = train_metadata.longitude.min()
lon_max = train_metadata.longitude.max()
train_metadata.longitude = (train_metadata.longitude - lon_min) / (lon_max - lon_min)
valid_metadata.longitude = (valid_metadata.longitude - lon_min) / (lon_max - lon_min)

wk_min = train_metadata.week.min()
wk_max = train_metadata.week.max()
train_metadata.week = (train_metadata.week - wk_min) / (wk_max - wk_min)
valid_metadata.week = (valid_metadata.week - wk_min) / (wk_max - wk_min)

dt_min = train_metadata.date_slno.min()
dt_max = train_metadata.date_slno.max()
train_metadata.date_slno = (train_metadata.date_slno - dt_min) / (dt_max - dt_min)
valid_metadata.date_slno = (valid_metadata.date_slno - dt_min) / (dt_max - dt_min)

valid_metadata.head(2)

train_metadata.head(2)

train_metadata = train_metadata.values
valid_metadata = valid_metadata.values

---

train_X.shape
train_y.shape
train_y_mask.shape
train_metadata.shape
valid_X.shape
valid_y.shape
valid_y_mask.shape
valid_metadata.shape

channel_means = train_X.reshape(-1, 128).mean(axis=0).reshape(1, 1, 1, -1)
channel_stds = train_X.reshape(-1, 128).std(axis=0).reshape(1, 1, 1, -1)
train_X = (train_X - channel_means) / channel_stds
valid_X = (valid_X - channel_means) / channel_stds
#np.save('data/channel_means.npy', channel_means)
#np.save('data/channel_stds.npy', channel_stds)

import gc
gc.collect()

# Prepare the PyTorch datasets

class AudioDataset(Dataset):

    def __init__(self, X, y, weights, metadata, transform=None):
        self.X = X
        self.y = y
        self.weights = weights
        self.metadata = metadata

    def __len__(self):
        return self.X.shape[0]

    def __getitem__(self, idx):
        sample = self.X[idx, ...]
        return sample, self.y[idx, ...], self.weights[idx, ...], self.metadata[idx, ...]

train_dataset = AudioDataset(torch.Tensor(train_X),
                             torch.Tensor(train_y),
                             torch.Tensor(train_y_mask),
                             torch.Tensor(train_metadata),
                             None)
valid_dataset = AudioDataset(torch.Tensor(valid_X),
                             torch.Tensor(valid_y),
                             torch.Tensor(valid_y_mask),
                             torch.Tensor(valid_metadata),
                             None)

val_loader = DataLoader(valid_dataset, 128, shuffle=False,
                        num_workers=2, drop_last=False, pin_memory=True)
train_loader_1 = DataLoader(train_dataset, 128, shuffle=True,
                            num_workers=2, drop_last=True, pin_memory=True)
train_loader_2 = DataLoader(train_dataset, 128, shuffle=True,
                            num_workers=2, drop_last=True, pin_memory=True)

# Define the model

cuda = True
device = torch.device('cuda:0' if cuda else 'cpu')
print('Device: ', device)

class Model(nn.Module):

    def __init__(self, num_classes):

        super().__init__()

        self.bw2col = nn.Sequential(
            nn.BatchNorm2d(1),
            nn.Conv2d(1, 10, 1, padding=0), nn.ReLU(),
            nn.Conv2d(10, 3, 1, padding=0), nn.ReLU())

        self.mv2 = torchvision.models.mobilenet_v2(pretrained=True)
        self.mv2.classifier = None
        
        self.after_mv2 = nn.Sequential(
            nn.BatchNorm1d(1280),
            nn.Dropout(0.7),
            nn.Linear(1280, 48),
            nn.ReLU(),
            nn.BatchNorm1d(48),
        )
        
        self.meta_seq = nn.Sequential(
            nn.BatchNorm1d(38),
            nn.Linear(38, 128),
            nn.ReLU(),
            nn.BatchNorm1d(128),
            nn.Linear(128, 16),
            nn.BatchNorm1d(16),
        )
        
        self.final_dense = nn.Linear(64, num_classes)

    def forward(self, x, with_metadata=False, metadata=None):
        
        x = self.bw2col(x)
        x = self.mv2.features(x)
        x = x.max(dim=-1)[0].max(dim=-1)[0]
        x = self.after_mv2(x)
        
        if not with_metadata:
            raise('this shouldnt happen')
        
        meta_out = self.meta_seq(metadata)
        
        x = torch.cat((x, meta_out), axis=-1)
        x = self.final_dense(x)
        return x

model = Model(31)

model

model = model.to(device)

# Bias training

optimizer = optim.Adam([model.final_dense.bias], lr=1, amsgrad=True)
scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience=5, verbose=True)
criterion = nn.BCEWithLogitsLoss(reduction='none')

num_bias_training = 2
train_loss_hist = []
valid_loss_hist = []
lowest_val_loss = np.inf
epochs_without_new_lowest = 0
train_batch_slno=0

for i in range(num_bias_training):
    print('Epoch: ', i)

    this_epoch_train_loss = 0
    for i1, i2 in zip(train_loader_1, train_loader_2):

        # mixup the inputs ---------
        alpha = 1
        mixup_vals = np.random.beta(alpha, alpha, i1[0].shape[0])

        lam = torch.Tensor(mixup_vals.reshape(mixup_vals.shape[0], 1, 1, 1))
        inputs = (lam * i1[0]) + ((1 - lam) * i2[0])

        lam = torch.Tensor(mixup_vals.reshape(mixup_vals.shape[0], 1, 1))
        labels = (lam * i1[1]) + ((1 - lam) * i2[1])
        masks = (lam * i1[2]) + ((1 - lam) * i2[2])
        
        lam = torch.Tensor(mixup_vals.reshape(mixup_vals.shape[0], 1))
        meta = (lam * i1[3]) + ((1 - lam) * i2[3])
        # mixup ends ----------

        inputs = inputs.to(device, non_blocking=True)
        labels = labels.to(device, non_blocking=True)
        masks = masks.to(device, non_blocking=True)
        meta = meta.to(device, non_blocking=True)

        optimizer.zero_grad()
        with torch.set_grad_enabled(True):
            model = model.train()
            outputs = model(inputs, with_metadata=True, metadata=meta)
            # calculate loss for each set of annotations
            loss_0 = criterion(outputs, labels[:, :, 0]) * masks[:, :, 0]
            loss_1 = criterion(outputs, labels[:, :, 1]) * masks[:, :, 1]
            loss_2 = criterion(outputs, labels[:, :, 2]) * masks[:, :, 2]
            loss = (loss_0.sum() + loss_1.sum() + loss_2.sum()) / masks.sum()
            loss.backward()
            optimizer.step()
            this_epoch_train_loss += loss.detach().cpu().numpy()
            tbwriter.add_scalar('Batch/Loss/Train',
                                loss.detach().cpu().numpy(),
                                train_batch_slno)
            train_batch_slno += 1

    this_epoch_valid_loss = 0
    for inputs, labels, masks, meta in val_loader:
        inputs = inputs.to(device)
        labels = labels.to(device)
        masks = masks.to(device)
        meta = meta.to(device)
        optimizer.zero_grad()
        with torch.set_grad_enabled(False):
            model = model.eval()
            outputs = model(inputs, with_metadata=True, metadata=meta)
            loss_0 = criterion(outputs, labels[:, :, 0]) * masks[:, :, 0]
            loss_1 = criterion(outputs, labels[:, :, 1]) * masks[:, :, 1]
            loss_2 = criterion(outputs, labels[:, :, 2]) * masks[:, :, 2]
            loss = (loss_0.sum() + loss_1.sum() + loss_2.sum()) / masks.sum()
            this_epoch_valid_loss += loss.detach().cpu().numpy()

    this_epoch_train_loss /= len(train_loader_1)
    this_epoch_valid_loss /= len(val_loader)

    train_loss_hist.append(this_epoch_train_loss)
    valid_loss_hist.append(this_epoch_valid_loss)

    if this_epoch_valid_loss < lowest_val_loss:
        lowest_val_loss = this_epoch_valid_loss
        torch.save(model.state_dict(), f"./models/{this_filename}")
        epochs_without_new_lowest = 0
    else:
        epochs_without_new_lowest += 1

    if epochs_without_new_lowest >= 25:
        break

    print(this_epoch_train_loss, this_epoch_valid_loss)
    tbwriter.add_scalar('Epoch/Loss/Train', this_epoch_train_loss, i)
    tbwriter.add_scalar('Epoch/Loss/Valid', this_epoch_valid_loss, i)
    tbwriter.add_scalar('Epoch/lr', optimizer.param_groups[0]['lr'], i)
    tbwriter.flush()
    
    scheduler.step(this_epoch_valid_loss)

# Training

optimizer = optim.Adam(model.parameters(), lr=0.0005, amsgrad=True)
scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience=5, verbose=True)
criterion = nn.BCEWithLogitsLoss(reduction='none')

epochs = 100
for i in range(num_bias_training, epochs):
    print('Epoch: ', i)

    this_epoch_train_loss = 0
    for i1, i2 in zip(train_loader_1, train_loader_2):

        # mixup the inputs ---------
        alpha = 1
        mixup_vals = np.random.beta(alpha, alpha, i1[0].shape[0])

        lam = torch.Tensor(mixup_vals.reshape(mixup_vals.shape[0], 1, 1, 1))
        inputs = (lam * i1[0]) + ((1 - lam) * i2[0])

        lam = torch.Tensor(mixup_vals.reshape(mixup_vals.shape[0], 1, 1))
        labels = (lam * i1[1]) + ((1 - lam) * i2[1])
        masks = (lam * i1[2]) + ((1 - lam) * i2[2])
        
        lam = torch.Tensor(mixup_vals.reshape(mixup_vals.shape[0], 1))
        meta = (lam * i1[3]) + ((1 - lam) * i2[3])
        # mixup ends ----------

        inputs = inputs.to(device, non_blocking=True)
        labels = labels.to(device, non_blocking=True)
        masks = masks.to(device, non_blocking=True)
        meta = meta.to(device, non_blocking=True)

        optimizer.zero_grad()
        with torch.set_grad_enabled(True):
            model = model.train()
            outputs = model(inputs, with_metadata=True, metadata=meta)
            # calculate loss for each set of annotations
            loss_0 = criterion(outputs, labels[:, :, 0]) * masks[:, :, 0]
            loss_1 = criterion(outputs, labels[:, :, 1]) * masks[:, :, 1]
            loss_2 = criterion(outputs, labels[:, :, 2]) * masks[:, :, 2]
            loss = (loss_0.sum() + loss_1.sum() + loss_2.sum()) / masks.sum()
            loss.backward()
            optimizer.step()
            this_epoch_train_loss += loss.detach().cpu().numpy()
            tbwriter.add_scalar('Batch/Loss/Train',
                                loss.detach().cpu().numpy(),
                                train_batch_slno)
            train_batch_slno += 1

    this_epoch_valid_loss = 0
    for inputs, labels, masks, meta in val_loader:
        inputs = inputs.to(device)
        labels = labels.to(device)
        masks = masks.to(device)
        meta = meta.to(device)
        optimizer.zero_grad()
        with torch.set_grad_enabled(False):
            model = model.eval()
            outputs = model(inputs, with_metadata=True, metadata=meta)
            loss_0 = criterion(outputs, labels[:, :, 0]) * masks[:, :, 0]
            loss_1 = criterion(outputs, labels[:, :, 1]) * masks[:, :, 1]
            loss_2 = criterion(outputs, labels[:, :, 2]) * masks[:, :, 2]
            loss = (loss_0.sum() + loss_1.sum() + loss_2.sum()) / masks.sum()
            this_epoch_valid_loss += loss.detach().cpu().numpy()

    this_epoch_train_loss /= len(train_loader_1)
    this_epoch_valid_loss /= len(val_loader)

    train_loss_hist.append(this_epoch_train_loss)
    valid_loss_hist.append(this_epoch_valid_loss)

    if this_epoch_valid_loss < lowest_val_loss:
        lowest_val_loss = this_epoch_valid_loss
        torch.save(model.state_dict(), f"./models/{this_filename}")
        epochs_without_new_lowest = 0
    else:
        epochs_without_new_lowest += 1

    if epochs_without_new_lowest >= 25:
        break

    print(this_epoch_train_loss, this_epoch_valid_loss)
    tbwriter.add_scalar('Epoch/Loss/Train', this_epoch_train_loss, i)
    tbwriter.add_scalar('Epoch/Loss/Valid', this_epoch_valid_loss, i)
    tbwriter.add_scalar('Epoch/lr', optimizer.param_groups[0]['lr'], i)
    tbwriter.flush()
    
    scheduler.step(this_epoch_valid_loss)

# Evaluation metrics

model.load_state_dict(torch.load(
    f"./models/{this_filename}"
))

this_metrics = {}

---

this_epoch_valid_loss = 0
val_preds = []
for inputs, labels, masks, meta in val_loader:
        inputs = inputs.to(device)
        labels = labels.to(device)
        masks = masks.to(device)
        meta = meta.to(device)
        optimizer.zero_grad()
        with torch.set_grad_enabled(False):
            model = model.eval()
            outputs = model(inputs, with_metadata=True, metadata=meta)
            loss_0 = criterion(outputs, labels[:, :, 0]) * masks[:, :, 0]
            loss_1 = criterion(outputs, labels[:, :, 1]) * masks[:, :, 1]
            loss_2 = criterion(outputs, labels[:, :, 2]) * masks[:, :, 2]
            loss = (loss_0.sum() + loss_1.sum() + loss_2.sum()) / masks.sum()
            this_epoch_valid_loss += loss.detach().cpu().numpy()
            val_preds.append(outputs.detach().cpu().numpy())
this_epoch_valid_loss /= len(val_loader)
val_preds = np.concatenate(val_preds, axis=0)

this_epoch_valid_loss, val_preds.shape

this_metrics['valid_loss'] = this_epoch_valid_loss

train_nodataaug_dataset = AudioDataset(torch.Tensor(train_X),
                                       torch.Tensor(train_y),
                                       torch.Tensor(train_y_mask),
                                       torch.Tensor(train_metadata),
                                       None)
train_nodataaug_loader = DataLoader(train_nodataaug_dataset, 64, shuffle=False)

this_epoch_train_nodataaug_loss = 0
train_preds = []
for inputs, labels, masks, meta in train_nodataaug_loader:
        inputs = inputs.to(device)
        labels = labels.to(device)
        masks = masks.to(device)
        meta = meta.to(device)
        optimizer.zero_grad()
        with torch.set_grad_enabled(False):
            model = model.eval()
            outputs = model(inputs, with_metadata=True, metadata=meta)
            loss_0 = criterion(outputs, labels[:, :, 0]) * masks[:, :, 0]
            loss_1 = criterion(outputs, labels[:, :, 1]) * masks[:, :, 1]
            loss_2 = criterion(outputs, labels[:, :, 2]) * masks[:, :, 2]
            loss = (loss_0.sum() + loss_1.sum() + loss_2.sum()) / masks.sum()
            this_epoch_train_nodataaug_loss += loss.detach().cpu().numpy()
            train_preds.append(outputs.detach().cpu().numpy())
this_epoch_train_nodataaug_loss /= len(train_nodataaug_loader)
train_preds = np.concatenate(train_preds, axis=0)

this_epoch_train_nodataaug_loss, train_preds.shape

this_metrics['train_loss'] = this_epoch_train_nodataaug_loss

---

our_cols = ['1_engine', '2_machinery-impact', '3_non-machinery-impact',
            '4_powered-saw', '5_alert-signal', '6_music', '7_human-voice', '8_dog',
            '1-1_small-sounding-engine', '1-2_medium-sounding-engine',
            '1-3_large-sounding-engine', '2-1_rock-drill', '2-2_jackhammer',
            '2-3_hoe-ram', '2-4_pile-driver', '3-1_non-machinery-impact',
            '4-1_chainsaw', '4-2_small-medium-rotating-saw',
            '4-3_large-rotating-saw', '5-1_car-horn', '5-2_car-alarm', '5-3_siren',
            '5-4_reverse-beeper', '6-1_stationary-music', '6-2_mobile-music',
            '6-3_ice-cream-truck', '7-1_person-or-small-group-talking',
            '7-2_person-or-small-group-shouting', '7-3_large-crowd',
            '7-4_amplified-speech', '8-1_dog-barking-whining']
cols_for_out = [
    "audio_filename", "1-1_small-sounding-engine",
    "1-2_medium-sounding-engine", "1-3_large-sounding-engine",
    "2-1_rock-drill",
    "2-2_jackhammer", "2-3_hoe-ram", "2-4_pile-driver",
    "3-1_non-machinery-impact",
    "4-1_chainsaw", "4-2_small-medium-rotating-saw",
    "4-3_large-rotating-saw",
    "5-1_car-horn", "5-2_car-alarm", "5-3_siren", "5-4_reverse-beeper",
    "6-1_stationary-music",
    "6-2_mobile-music", "6-3_ice-cream-truck",
    "7-1_person-or-small-group-talking",
    "7-2_person-or-small-group-shouting", "7-3_large-crowd",
    "7-4_amplified-speech",
    "8-1_dog-barking-whining", "1_engine", "2_machinery-impact",
    "3_non-machinery-impact", "4_powered-saw", "5_alert-signal",
    "6_music", "7_human-voice", "8_dog"]

---

val_df = pd.DataFrame((1 / (1 + np.exp(-val_preds))), columns = our_cols)
val_df['audio_filename'] = pd.Series(
    sorted(set(metadata['coarse_valid'].index.tolist())),
    index=val_df.index)
val_df = val_df.loc[:, cols_for_out]
val_df = val_df.loc[lambda x: x.audio_filename.isin(
    metadata['coarse_valid_gt'].index.tolist())]

val_df.to_csv('/tmp/val_subm.csv', index=False)

---

import sys
sys.path.append('./baseline_code')

from metrics import evaluate, macro_averaged_auprc, micro_averaged_auprc

---

df_dict = evaluate(
    '/tmp/val_subm.csv',
    './data/annotations.csv',
    './data/dcase-ust-taxonomy.yaml',
    'coarse'
)
micro_auprc, eval_df = micro_averaged_auprc(df_dict, return_df=True)
macro_auprc, class_auprc = macro_averaged_auprc(df_dict, return_classwise=True)
thresh_0pt5_idx = (eval_df['threshold'] >= 0.5).to_numpy().nonzero()[0][0]

micro_auprc, macro_auprc, eval_df['F'][thresh_0pt5_idx]
class_auprc

this_metrics['valid_coarse_micro_auprc'] = micro_auprc
this_metrics['valid_coarse_macro_auprc'] = macro_auprc
this_metrics['valid_coarse_f1'] = eval_df['F'][thresh_0pt5_idx]
for k, v in class_auprc.items():
    this_metrics[f"valid_coarse_auprc_class_{k}"] = v

df_dict = evaluate(
    '/tmp/val_subm.csv',
    './data/annotations.csv',
    './data/dcase-ust-taxonomy.yaml',
    'fine'
)
micro_auprc, eval_df = micro_averaged_auprc(df_dict, return_df=True)
macro_auprc, class_auprc = macro_averaged_auprc(df_dict, return_classwise=True)
thresh_0pt5_idx = (eval_df['threshold'] >= 0.5).to_numpy().nonzero()[0][0]

micro_auprc, macro_auprc, eval_df['F'][thresh_0pt5_idx]
class_auprc

this_metrics['valid_fine_micro_auprc'] = micro_auprc
this_metrics['valid_fine_macro_auprc'] = macro_auprc
this_metrics['valid_fine_f1'] = eval_df['F'][thresh_0pt5_idx]
for k, v in class_auprc.items():
    this_metrics[f"valid_fine_auprc_class_{k}"] = v

---

train_nodataaug_df = pd.DataFrame((1 / (1 + np.exp(-train_preds))), columns = our_cols)
train_nodataaug_df['audio_filename'] = pd.Series(
    sorted(set(metadata['coarse_train'].index.tolist())),
    index=train_nodataaug_df.index)
train_nodataaug_df = train_nodataaug_df.loc[:, cols_for_out]
train_nodataaug_df = train_nodataaug_df.loc[lambda x: x.audio_filename.isin(
    metadata['coarse_train_gt'].index.tolist())]

train_nodataaug_df.to_csv('/tmp/train_nodataaug.csv', index=False)

df_dict = evaluate(
    '/tmp/train_nodataaug.csv',
    './data/annotations.csv',
    './data/dcase-ust-taxonomy.yaml',
    'coarse',
    True
)
micro_auprc, eval_df = micro_averaged_auprc(df_dict, return_df=True)
macro_auprc, class_auprc = macro_averaged_auprc(df_dict, return_classwise=True)
thresh_0pt5_idx = (eval_df['threshold'] >= 0.5).to_numpy().nonzero()[0][0]

micro_auprc, macro_auprc, eval_df['F'][thresh_0pt5_idx]
class_auprc

this_metrics['train_coarse_micro_auprc'] = micro_auprc
this_metrics['train_coarse_macro_auprc'] = macro_auprc
this_metrics['train_coarse_f1'] = eval_df['F'][thresh_0pt5_idx]
for k, v in class_auprc.items():
    this_metrics[f"train_coarse_auprc_class_{k}"] = v

df_dict = evaluate(
    '/tmp/train_nodataaug.csv',
    './data/annotations.csv',
    './data/dcase-ust-taxonomy.yaml',
    'fine',
    True
)
micro_auprc, eval_df = micro_averaged_auprc(df_dict, return_df=True)
macro_auprc, class_auprc = macro_averaged_auprc(df_dict, return_classwise=True)
thresh_0pt5_idx = (eval_df['threshold'] >= 0.5).to_numpy().nonzero()[0][0]

micro_auprc, macro_auprc, eval_df['F'][thresh_0pt5_idx]
class_auprc

this_metrics['train_fine_micro_auprc'] = micro_auprc
this_metrics['train_fine_macro_auprc'] = macro_auprc
this_metrics['train_fine_f1'] = eval_df['F'][thresh_0pt5_idx]
for k, v in class_auprc.items():
    this_metrics[f"train_fine_auprc_class_{k}"] = v

---

this_metrics = {
    f'hparams/{k}': v
    for k, v in this_metrics.items()
}

tbwriter.add_hparams(hparam_dict=this_hparams, metric_dict=this_metrics)

tbwriter.flush()
